pyproject.toml
[tool.poetry]
name = "derby"
version = "1.0.0"
description = "a bidding, auction, and market framework for creating and running auction or market games."
authors = ["Nishant Kumar <nish.ku.121@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26"
pandas = "^1.5"
matplotlib = "^3.7"
scipy = "^1.11"
scikit-learn = "^1.3"
tensorflow = "2.16.1"
tensorflow-probability = {version = "0.24.0", extras = ["tf"]}
protobuf = "^3.20"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
notebook = "^6.5.7"
jupyter = "^1.1.1"
ipykernel = "^6.28"
ipywidgets = "^8.0.6"
qtconsole = "^5.4.2"
nbconvert = "^7.7.0"
nbformat = "^5.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
